+++ source/ftrack_api/event/hub.py
@@ -186,17 +186,26 @@ class EventHub(object):
             # https://docs.python.org/2/library/socket.html#socket.socket.setblocking
             self._connection = websocket.create_connection(url, timeout=60)
 
-        except Exception:
+        except Exception as error:
+            error_message = (
+                'Failed to connect to event server at {server_url} with '
+                'error: "{error}".'
+            )
+
+            error_details = {
+                'error': unicode(error),
+                'server_url': self.get_server_url()
+            }
+
             self.logger.debug(
                 L(
-                    'Error connecting to event server at {0}.',
-                    self.get_server_url()
+                    error_message, **error_details
                 ),
                 exc_info=1
             )
             raise ftrack_api.exception.EventHubConnectionError(
-                'Failed to connect to event server at {0}.'
-                .format(self.get_server_url())
+                error_message,
+                details=error_details
             )
 
         # Start background processing thread.
@@ -913,6 +922,7 @@ class EventHub(object):
         if code_name == 'connect':
             self.logger.debug('Connected to event server.')
             event = ftrack_api.event.base.Event('ftrack.meta.connected')
+            self._prepare_event(event)
             self._event_queue.put(event)
 
         elif code_name == 'disconnect':
@@ -933,6 +943,7 @@ class EventHub(object):
 
             if not self.connected:
                 event = ftrack_api.event.base.Event('ftrack.meta.disconnected')
+                self._prepare_event(event)
                 self._event_queue.put(event)
 
         elif code_name == 'heartbeat':