+++ source/ftrack_api/logging.py
@@ -2,7 +2,22 @@
 # :copyright: Copyright (c) 2016 ftrack
 
 from builtins import object
+import functools
+import warnings
 
+
+def deprecation_warning(message):
+    def decorator(function):
+        @functools.wraps(function)
+        def wrapper(*args, **kwargs):
+            warnings.warn(
+                message,
+                PendingDeprecationWarning
+            )
+            return function(*args, **kwargs)
+        return wrapper
+
+    return decorator
 
 class LazyLogMessage(object):
     '''A log message that can be evaluated lazily for improved performance.